Easy - Pascal triangle
The problem of generating Pascal's triangle can be approached in various ways.
Approach : Using Recursion
Intuition: In Pascal's triangle, each element is the sum of the two elements directly above it. We can use a recursive approach to generate the triangle. The base case is when numRows is 1, in which case we return [[1]]. Otherwise, we recursively generate the triangle for numRows - 1 and then calculate the current row by summing the adjacent elements from the previous row.