Rename Medium - to Medium3 - Maximal Square
Approach
This problem can be solved using dynamic programming. We can define a 2D array dp where dp[i][j] represents the maximum size of a square that can be formed at position (i, j) such that all its elements are 1's. We can fill this array using the following recurrence relation:

dp[i][j] = 0 if matrix[i][j] == '0'
dp[i][j] = min(dp[i-1][j], dp[i][j-1], dp[i-1][j-1]) + 1 if matrix[i][j] == '1'
The first condition means that if the current cell has a value of 0, we can't form a square with it. The second condition means that if the current cell has a value of 1, we can form a square with it, but we need to check the values of the cells to the left, top, and top-left of it to determine the maximum size of the square.

We also need to keep track of the maximum size of the square seen so far, and return its area.

Complexity
Time complexity:
Beats
88.37% O(m*n), where m and n are the dimensions of the matrix.

Space complexity:
Beats
88.43% O(m*n), where m and n are the dimensions of the matrix, since we are using a 2D array of the same size as the matrix to store the dp values.

The algorithm iterates through each cell of the matrix, filling the dp array with the sizes of squares ending at each cell.
The maximum size of the square seen so far is continuously updated.
The result is the area of the largest square found, which is the square of the maximum size.